v1.10.0

+++ src/supabase/profiles.ts
@@ -1,4 +1,4 @@
-import { SUPABASE } from '../helpers';
+import { SUPABASE } from './supabaseClient';
 import Logger from '../logger';
 import { Profile } from '../types/supabase';
 
@@ -11,6 +11,7 @@ const LOG = new Logger('profiles');
  * @returns {Promise<Profile | null>} | User data
  */
 export const getProfileById = async (id: string): Promise<Profile | null> => {
+    if (!SUPABASE) return null;
     try {
         const { data, error } = await SUPABASE.from('profiles').select().eq('id', id).single();
 
@@ -36,6 +37,7 @@ export const updateProfileUsername = async (
     id: string,
     username: string
 ): Promise<Profile | null> => {
+    if (!SUPABASE) return null;
     try {
         const { data, error } = await SUPABASE.from('profiles')
             .update({ username: username })
@@ -61,6 +63,7 @@ export const updateProfileUsername = async (
  * @returns {Promise<void>}
  */
 export const deleteProfileById = async (id: string): Promise<void> => {
+    if (!SUPABASE) return;
     try {
         // delete profile
         const { data, error } = await SUPABASE.from('profiles').delete().eq('id', id).select();
@@ -88,14 +91,20 @@ export const deleteProfileById = async (id: string): Promise<void> => {
  * @param id | uuid users profile being queried
  * @returns {Promise<number[] | null>}
  */
-export const getProfileQueue = async (id: string): Promise<string[] | null> => {
+export const getProfileArray = async (
+    profileId: string,
+    whichCol: 'queue' | 'watched' | 'favorites'
+): Promise<string[] | null> => {
+    if (!SUPABASE) return null;
     try {
-        const { data, error } = await SUPABASE.from('profiles').select('queue').eq('id', id);
+        const { data, error } = await SUPABASE.from('profiles')
+            .select(whichCol)
+            .eq('id', profileId);
 
         if (error) {
             LOG.error(error);
         } else if (data) {
-            return data[0].queue;
+            return Object.values(data[0])[0];
         }
     } catch (error) {
         LOG.error(error as string);
@@ -112,15 +121,16 @@ export const getProfileQueue = async (id: string): Promise<string[] | null> => {
  * @returns {Promise<Profile | null>}
  */
 export const addToProfileArray = async (
-    profile_id: string,
-    show_id: string,
-    which_col: 'queue' | 'watched' | 'favorites'
+    profileId: string,
+    showId: string,
+    whichCol: 'queue' | 'watched' | 'favorites'
 ): Promise<Profile | null> => {
+    if (!SUPABASE) return null;
     try {
         const { data, error } = await SUPABASE.rpc('add_item', {
-            profile_id,
-            show_id,
-            which_col,
+            profile_id: profileId,
+            show_id: showId,
+            which_col: whichCol,
         })
             .select()
             .single();
@@ -145,15 +155,16 @@ export const addToProfileArray = async (
  * @returns {Promise<Profile | null>}
  */
 export const removeFromProfileArray = async (
-    profile_id: string,
-    show_id: string,
-    which_col: 'queue' | 'watched' | 'favorites'
+    profileId: string,
+    showId: string,
+    whichCol: 'queue' | 'watched' | 'favorites'
 ): Promise<Profile | null> => {
+    if (!SUPABASE) return null;
     try {
         const { data, error } = await SUPABASE.rpc('remove_item', {
-            profile_id,
-            show_id,
-            which_col,
+            profile_id: profileId,
+            show_id: showId,
+            which_col: whichCol,
         })
             .select()
             .single();
@@ -173,17 +184,18 @@ export const removeFromProfileArray = async (
  * Removes all shows from a users queue, watched, or favorites
  *
  * @param id | uuid of user being updated
- * @param which_col | profiles column to delete
+ * @param whichCol | profiles column to delete
  * @returns {Promise<Profile | null>}
  */
-export const removeProfileArray = async (
-    profile_id: string,
-    which_col: 'queue' | 'watched' | 'favorites'
+export const clearProfileArray = async (
+    profileId: string,
+    whichCol: 'queue' | 'watched' | 'favorites'
 ): Promise<Profile | null> => {
+    if (!SUPABASE) return null;
     try {
         const { data, error } = await SUPABASE.rpc('remove_all', {
-            profile_id,
-            which_col,
+            profile_id: profileId,
+            which_col: whichCol,
         })
             .select()
             .single();
@@ -210,6 +222,7 @@ export const setProfileAdultFlag = async (
     id: string,
     isAdult: boolean
 ): Promise<Profile | null> => {
+    if (!SUPABASE) return null;
     try {
         const { data, error } = await SUPABASE.from('profiles')
             .update({ adult: isAdult })
@@ -235,6 +248,7 @@ export const setProfileAdultFlag = async (
  * @returns {Promise<Profile | null>}
  */
 export const setProfileCountry = async (id: string, country: string): Promise<Profile | null> => {
+    if (!SUPABASE) return null;
     try {
         // TODO: #587 Ensure country code is valid
         const { data, error } = await SUPABASE.from('profiles')
